Revert "refactor: main에서 props 분리"
This reverts commit 7e8becf11114c107c4cc0e111029696ed537c834.
