+++ src/pages/Main/Main.tsx
@@ -8,8 +8,145 @@ import { Mixpanel, TRACK } from '../../common/mixpanel';
 import CardListView from '../../views/CardList';
 import logoImage from '../../img/logo.png';
 import donationLogo from '../../img/donation.svg';
-import GlobalChoiceDialog from '../../components/GeneralChoiceDialog';
-import { selectedCardReducer, resetState } from './Props';
+import GlobalChoiceDialog from "../../components/GeneralChoiceDialog";
+
+type SetUmaAction = {
+  type: 'SET_UMA_ACTION';
+  payload: number;
+};
+
+type AppendAction = {
+  type: 'APPEND_CARD';
+  payload: number;
+};
+
+type DeleteAction = {
+  type: 'DELETE_CARD';
+  payload: number;
+};
+
+type SetCardAction = {
+  type: 'SET_CARD';
+  payload: number;
+};
+
+type ToggleUmaPageAction = {
+  type: 'TOGGLE_UMA_PAGE';
+};
+
+type ToggleCardPageAction = {
+  type: 'TOGGLE_CARD_PAGE';
+};
+
+type ResetFavoriteAction = {
+  type: 'RESET_FAVORITE';
+};
+
+type ToggleGeneralChoice = {
+  type: 'TOGGLE_GENERAL_CHOICE';
+}
+
+const resetState = (): State => {
+  const favoritesString = localStorage.getItem('favoriteCardUuids');
+  return {
+    favoriteCardUuids: favoritesString ? JSON.parse(favoritesString) : [],
+    cardUuid: 0,
+    umaUuid: 0,
+    showUmaPage: true,
+    showCardPage: true,
+    showGeneralChoice: false,
+  };
+};
+
+type Action =
+  | AppendAction
+  | DeleteAction
+  | ResetFavoriteAction
+  | SetUmaAction
+  | SetCardAction
+  | ToggleUmaPageAction
+  | ToggleCardPageAction
+  | ToggleGeneralChoice;
+
+type State = {
+  favoriteCardUuids: number[];
+  umaUuid: number;
+  cardUuid: number;
+  showUmaPage: boolean;
+  showCardPage: boolean;
+  showGeneralChoice: boolean;
+};
+
+const selectedCardReducer = (state: State, action: Action): State => {
+  switch (action.type) {
+    case 'SET_UMA_ACTION':
+      return {
+        ...state,
+        umaUuid: action.payload,
+        showUmaPage: !state.showUmaPage,
+      };
+    case 'APPEND_CARD': {
+      if (state.favoriteCardUuids.includes(action.payload)) {
+        const newFavorites = state.favoriteCardUuids.filter(
+          (uuid) => uuid !== action.payload
+        );
+        localStorage.setItem('favoriteCardUuids', JSON.stringify(newFavorites));
+        return {
+          ...state,
+          favoriteCardUuids: newFavorites,
+        };
+      }
+      if (state.favoriteCardUuids.length >= 6) {
+        return state;
+      }
+      const newFavorites = [...state.favoriteCardUuids, action.payload];
+      localStorage.setItem('favoriteCardUuids', JSON.stringify(newFavorites));
+      return {
+        ...state,
+        favoriteCardUuids: newFavorites,
+      };
+    }
+    case 'DELETE_CARD': {
+      const newFavorites = state.favoriteCardUuids.filter(
+        (x) => x !== action.payload
+      );
+      localStorage.setItem('favoriteCardUuids', JSON.stringify(newFavorites));
+      return {
+        ...state,
+        favoriteCardUuids: newFavorites,
+      };
+    }
+    case 'SET_CARD':
+      return {
+        ...state,
+        cardUuid: action.payload,
+        showCardPage: false,
+      };
+    case 'RESET_FAVORITE':
+      localStorage.setItem('favoriteCardUuids', JSON.stringify([]));
+      return {
+        ...state,
+        favoriteCardUuids: [],
+      };
+    case 'TOGGLE_UMA_PAGE':
+      return {
+        ...state,
+        showUmaPage: !state.showUmaPage,
+      };
+    case 'TOGGLE_CARD_PAGE':
+      return {
+        ...state,
+        showCardPage: !state.showCardPage,
+      };
+    case 'TOGGLE_GENERAL_CHOICE':
+      return {
+        ...state,
+        showGeneralChoice: !state.showGeneralChoice
+      }
+    default:
+      throw new Error('Unhandled action');
+  }
+};
 
 const Main: React.FC = () => {
   const [state, dispatch] = useReducer(selectedCardReducer, {}, resetState);
@@ -31,7 +168,7 @@ const Main: React.FC = () => {
   };
   const toggleUmaPage = () => dispatch({ type: 'TOGGLE_UMA_PAGE' });
   const toggleCardPage = () => dispatch({ type: 'TOGGLE_CARD_PAGE' });
-  const toggleGeneralChoice = () => dispatch({ type: 'TOGGLE_GENERAL_CHOICE' });
+  const toggleGeneralChoice = () => dispatch({ type: 'TOGGLE_GENERAL_CHOICE' })
 
   useEffect(() => {
     Mixpanel.track(TRACK.MAINPAGE, {});
@@ -49,13 +186,8 @@ const Main: React.FC = () => {
         {/* <div>우마서포터</div> */}
         <span className={'TitleBarRightItems'}>
           <a className={'report'} href={'https://forms.gle/wkMFKgKyb2839GRL6'}>
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              id="mdi-alarm-light-outline"
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-            >
+            <svg xmlns="http://www.w3.org/2000/svg"
+              id="mdi-alarm-light-outline" width="24" height="24" viewBox="0 0 24 24">
               <title>정보 제보하기</title>
               <path d="M6,6.9L3.87,4.78L5.28,3.37L7.4,5.5L6,6.9M13,1V4H11V1H13M20.13,4.78L18,6.9L16.6,5.5L18.72,3.37L20.13,4.78M4.5,10.5V12.5H1.5V10.5H4.5M19.5,10.5H22.5V12.5H19.5V10.5M6,20H18A2,2 0 0,1 20,22H4A2,2 0 0,1 6,20M12,5A6,6 0 0,1 18,11V19H6V11A6,6 0 0,1 12,5M12,7A4,4 0 0,0 8,11V17H16V11A4,4 0 0,0 12,7Z" />
             </svg>
@@ -104,9 +236,8 @@ const Main: React.FC = () => {
             .reduce((a, b) => a + b)}`}
         >
           <div
-            className={`UmaEventChoice EventChoice ${
-              !state.showUmaPage ? 'activated' : ''
-            }`}
+            className={`UmaEventChoice EventChoice ${!state.showUmaPage ? 'activated' : ''
+              }`}
           >
             <UmaDetailContainer
               uuid={state.umaUuid}
@@ -115,9 +246,8 @@ const Main: React.FC = () => {
           </div>
 
           <div
-            className={`CardEventChoice EventChoice ${
-              !state.showCardPage ? 'activated' : ''
-            }`}
+            className={`CardEventChoice EventChoice ${!state.showCardPage ? 'activated' : ''
+              }`}
           >
             <SupportCardDetailContainer
               uuid={state.cardUuid}
@@ -138,16 +268,13 @@ const Main: React.FC = () => {
           />
         </div>
       </div>
-      <div className={'GlobalChoice'}>
-        <div className={'GlobalChoiceButton'} onClick={toggleGeneralChoice}>
-          우마무스메 / 서포터 외의 이벤트
-        </div>
+      <div className={"GlobalChoice"}>
+        <div className={"GlobalChoiceButton"} onClick={toggleGeneralChoice}>우마무스메 / 서포터 외의 이벤트</div>
+
       </div>
-      <GlobalChoiceDialog
-        open={state.showGeneralChoice}
-        onClose={toggleGeneralChoice}
-      ></GlobalChoiceDialog>
-    </Div100vh>
+      <GlobalChoiceDialog open={state.showGeneralChoice} onClose={toggleGeneralChoice}>
+      </GlobalChoiceDialog>
+    </Div100vh >
   );
 };
